fix(test): deduplicate lwd_integration_test and restore missing doc comment
Move the 265-line lwd_integration_test helper to common/lightwalletd.rs
(its logical home alongside spawn_lightwalletd_for_rpc) instead of
duplicating it in both stateful/sync.rs and integration/network.rs.

Also restore the 2 missing doc comment lines on non_blocking_logger
that were dropped during the initial extraction.
